% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{AddSentiStrengthData}
\alias{AddSentiStrengthData}
\title{Add SentiStrength Data}
\usage{
AddSentiStrengthData(location, datadir = "sentidata_en", overwrite = FALSE)
}
\arguments{
\item{location}{Location of the folder data.}

\item{datadir}{Name of the destination folder.}

\item{overwrite}{If FALSE, the function will fail if the file
already exists.}
}
\description{
Specifies SentiStrength language data and copies it inside the
installed package folder.
}
