% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentistrength.R
\name{Polarity}
\alias{Polarity}
\title{Polarity}
\usage{
Polarity(positive, negative, neutral.limit = 4)
}
\arguments{
\item{positive}{Positive scores as computed by SentiStrength}

\item{negative}{Negative scores as computed by SentiStrength}

\item{neutral.limit}{If positive and negative are equals and are
below this below, then the result is neutral, otherwise NA.}
}
\value{
-1, 0, 1 or NA based on the overall computed polarity.
}
\description{
Computes the polarity of a text document based on SentiStrength's
positive and negative scores.
}
\examples{
Polarity(1, -1)
Polarity(2, -1)
Polarity(1, -2)
Polarity(3, -3)
Polarity(c(5, 4, 4), c(-4, -5, -4))
Polarity(5, -5, 6)
Polarity(c(1, 2), c(-1, -2), 2)
}
\references{
M. Thelwall, K. Buckley, and G. Paltoglou,
  "Sentiment strength detection for the social web,"
  *J. Am. Soc. Inf. Sci. Technol.*, vol. 63, no. 1, pp. 163–173,
  Jan. 2012.
}
