% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentistrength.R
\name{SentiStrength}
\alias{SentiStrength}
\title{SentiStrength}
\usage{
SentiStrength(
  text,
  sentistrength.data = SentiStrengthData("sentidata_en"),
  parse = TRUE,
  additional.args = c()
)
}
\arguments{
\item{text}{The text on which to run SentiStrength.}

\item{sentistrength.data}{Location of the folder containing the
SentiStrength language data.}

\item{parse}{If TRUE, parses the output of SentiStrength into a
\code{data.table} object.}

\item{additional.args}{Additional argument to pass to the Java CLI
of SentiStrength.}
}
\value{
The output of SentiStrength as a character vector or parsed
  as a \code{data.table} object.
}
\description{
Runs SentiStrength
}
