% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SentiStrengthData}
\alias{SentiStrengthData}
\title{SentiStrength Data}
\usage{
SentiStrengthData(datadir)
}
\arguments{
\item{datadir}{Name of the folder where the language data is stored
inside the installed package.}
}
\value{
The path to the data folder. Fails if it doesn't exist.
}
\description{
Returns the location of a specific language data.
}
\seealso{
AddSentiStrengthData
}
