% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SentiStrengthJar}
\alias{SentiStrengthJar}
\title{SentiStrength Jar}
\usage{
SentiStrengthJar(location, overwrite = FALSE)
}
\arguments{
\item{location}{Location of the file. Can be a local file name or a
URL.}

\item{overwrite}{If FALSE, the function will fail if the file
already exists.}
}
\description{
Specifies SentiStrength jar and copies it inside the installed
package folder.
}
